Started refactoring changes to debug code
