--- Assets/Samples/SimplifiedDemoKCC/Editor/DrawKCCGrounded.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bc83e96bf8f7d554b84c678e78556a9e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
