--- Assets/Samples/SimplifiedDemoKCC/Editor/DrawKCCGroundedAngle.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: fdd1547483fc3cd439e2ebcd64f7e2fc
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
