--- Assets/Samples/SimplifiedDemoKCC/Editor/DrawKCCSteps.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 68893529ef92f984298e053cda69e92e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
