+++ Assets/Samples/SimplifiedDemoKCC/EditorGizmos/ColliderCastExtensions.cs
@@ -20,7 +20,7 @@ using nickmaltbie.OpenKCC.Utils;
 using nickmaltbie.OpenKCC.Utils.ColliderCast;
 using UnityEngine;
 
-namespace nickmaltbie.OpenKCC.Editor
+namespace nickmaltbie.OpenKCC.Demo.Debug
 {
     /// <summary>
     /// Collider cast extensions for debug drawing.
