+++ Assets/Samples/SimplifiedDemoKCC/Scripts/DrawKCCBounces.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b7a246478bc699441a17bf2dfdc04a0e
+guid: 4efc0b7acc0c8a74f9f561a2f713a1f1
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
