+++ Assets/Samples/SimplifiedDemoKCC/SimplifiedScene.unity
@@ -20094,7 +20094,7 @@ Transform:
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 1
+  m_RootOrder: 2
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &998168075
 GameObject:
@@ -20164,7 +20164,7 @@ Transform:
   - {fileID: 572539590}
   - {fileID: 1167349255}
   m_Father: {fileID: 0}
-  m_RootOrder: 6
+  m_RootOrder: 7
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &1008670904
 GameObject:
@@ -41756,6 +41756,63 @@ Mesh:
     offset: 0
     size: 0
     path: 
+--- !u!1001 &1693173875
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 640097779924159904, guid: bb557db3b565cd346a492ee35159a161, type: 3}
+      propertyPath: m_Name
+      value: SimplifiedPlayer (1)
+      objectReference: {fileID: 0}
+    - target: {fileID: 640097779924159911, guid: bb557db3b565cd346a492ee35159a161, type: 3}
+      propertyPath: m_RootOrder
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 640097779924159911, guid: bb557db3b565cd346a492ee35159a161, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0.14435863
+      objectReference: {fileID: 0}
+    - target: {fileID: 640097779924159911, guid: bb557db3b565cd346a492ee35159a161, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: -0.0000009536743
+      objectReference: {fileID: 0}
+    - target: {fileID: 640097779924159911, guid: bb557db3b565cd346a492ee35159a161, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 27.606815
+      objectReference: {fileID: 0}
+    - target: {fileID: 640097779924159911, guid: bb557db3b565cd346a492ee35159a161, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 640097779924159911, guid: bb557db3b565cd346a492ee35159a161, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 640097779924159911, guid: bb557db3b565cd346a492ee35159a161, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 640097779924159911, guid: bb557db3b565cd346a492ee35159a161, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 640097779924159911, guid: bb557db3b565cd346a492ee35159a161, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 640097779924159911, guid: bb557db3b565cd346a492ee35159a161, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 640097779924159911, guid: bb557db3b565cd346a492ee35159a161, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: bb557db3b565cd346a492ee35159a161, type: 3}
 --- !u!1 &1711891617
 GameObject:
   m_ObjectHideFlags: 0
@@ -51928,7 +51985,7 @@ Transform:
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 2
+  m_RootOrder: 3
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!114 &1946934080
 MonoBehaviour:
