--- Packages/com.nickmaltbie.openkcc/Editor.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 4f3e4d6699edbee42b4a6b9385833ada
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
