--- Packages/com.nickmaltbie.openkcc/Editor/nickmaltbie.OpenKCC.Editor.asmdef
@@ -1,18 +1,0 @@
-{
-    "name": "nickmaltbie.OpenKCC.Editor",
-    "rootNamespace": "nickmaltbie.OpenKCC.Editor",
-    "references": [
-        "GUID:a39c454983160b74bad1ea1398c0807f"
-    ],
-    "includePlatforms": [
-        "Editor"
-    ],
-    "excludePlatforms": [],
-    "allowUnsafeCode": false,
-    "overrideReferences": false,
-    "precompiledReferences": [],
-    "autoReferenced": true,
-    "defineConstraints": [],
-    "versionDefines": [],
-    "noEngineReferences": false
-}
