--- Packages/com.nickmaltbie.openkcc/Editor/nickmaltbie.OpenKCC.Editor.asmdef.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: e6fea7572fbe13649b61cd6138fab80e
-AssemblyDefinitionImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
